extract world panel
