+++ crates/client/src/ui/panels/mod.rs
@@ -6,6 +6,7 @@ pub mod inventory;
 pub mod look;
 pub mod menu;
 pub mod quit;
+pub mod world;
 
 pub use body::render_body_panel;
 pub use crafting::render_crafting_panel;
@@ -15,3 +16,4 @@ pub use inventory::{get_player_inventory, render_inventory_list_only, render_inv
 pub use look::render_look_panel;
 pub use menu::render_menu_panel;
 pub use quit::render_quit_panel;
+pub use world::render_game_view;
